move the tests that require reverse proxy settings to a special package
